Add files via upload
This program is to solve the cannibals and missionaries problem for different situations. 
Basically the problem is that the number of cannibals cannot be larger than the number of missionaries at the both sides of the river and the boat. The program tries to find a possible way to carry the missionaries and the cannibals across the river. 
This program is for:
a. 5 cannibals, 5 missionaries and a boat holding up to 3
b. 6 cannibals, 6 missionaries and a boat holding up to 4