Add files via upload
Solving a variant of the 15-puzzle, namely E15 puzzle by using A* algorithm. 
The program creates 12 random state space with shuffling the goal state by 10 moves. After the random space state is created, the program uses A* algorithm to reach the goal state which is: goal_state = [[1,2,3,4],
              [2,3,4,5],
              [3,4,5,5],
              [4,5,5," "]]   
For the first 2 solution, the move of tiles are shown. 
For the rest(10solution), the length of the queue is shown by a plot.