Add files via upload
This program does minimaxing and alphabeta pruning for a tree having a depth of 2 and having a branch of 3. Hence, at the leafs; there are nine nodes connected to 3 middle nodes (depth=1). These three nodes are connected to root node(depth=0). The name of the nodes at leafs are A, B, C, D, E, F, G, H, I. 
The program does the minimaxing for 2 cases: 
a. The leaf nodes are [5,3,1,2,5,4,1,3,3]. 
b. The leaf nodes are created customly by the user.
The program does the alpha-beta pruning for 4 cases. 
a. The leaf nodes are [5,3,1,2,5,4,1,3,3].
b. The leaf nodes are [5,2,2,5,1,3,2,4,2].
c. The leaf nodes are [1,3,4,1,4,1,3,5,3].
d. The leaf nodes are created customly by the user.